Extract TapeGuard_ to shared curve header
- Add dal-cpp/dal/curve/tapeguard.hpp with the RAII TapeGuard_ struct
  (Clear on construct/destruct, copy and assign deleted)
- Remove the two byte-identical anonymous-namespace TapeGuard_ definitions
  from calibration.cpp and jointcalibration.cpp; include the new header
  from both, placed after <dal/math/aad/aad.hpp>
- Call sites (TapeGuard_ guard(tape);) compile unchanged via the shared
  namespace-Dal definition

The two anonymous-namespace copies had diverged in surrounding comments
and would diverge in implementation under any future change. Promoting
to a single header makes the ODR contract machine-enforceable and gives
both single-curve and joint analytic-Jacobian paths one definition.
